fix(general): reorganised my structure
