+++ src/app/features/blog-details-page/blog-details-page.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BlogDetailsPageComponent } from './blog-details-page.component';
-import { BlogItemComponent } from '@app/core/components/blog-overview/blog-item';
+import { BlogItemComponent } from '@app/shared/blog-overview/blog-item';
 import { HttpClientModule } from '@angular/common/http';
 import { MatCardModule } from '@angular/material/card';
 import { BlogService } from '@app/services/blog.service';
 import { Routes, RouterModule } from '@angular/router';
-import { CommentItemComponent } from '@app/core/comment-item/comment-item';
+import { CommentItemComponent } from '@app/shared/comment-item/comment-item';
 
 const routes: Routes = [
   {
