+++ src/app/features/blog-overview-page/blog-overview-page.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BlogOverviewPageComponent } from '@features/blog-overview-page/blog-overview-page.component';
-import { BlogItemComponent } from '@app/core/components/blog-overview/blog-item';
+import { BlogItemComponent } from '@app/shared/blog-overview/blog-item';
 import { BlogService } from '@app/services/blog.service';
 import { HttpClientModule } from '@angular/common/http';
 import { MatCardModule } from '@angular/material/card';
