+++ src/app/shared/blog-overview/blog-item.component.scss
