+++ src/app/shared/comment-item/comment-item.component.scss
